requirements.txt
fastapi
httpx[socks]
openai
python-dotenv
rich
starlette
win_inet_pton
